% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility-internal.R
\name{check_class}
\alias{check_class}
\title{Check parameter class}
\usage{
check_class(param, expect_class)
}
\description{
Check the class of a supplied parameter to make sure it matches
expectations.
}
\keyword{internal}

