% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_base_url-internal.R
\name{get_base_url}
\alias{get_base_url}
\title{get_base_url}
\usage{
get_base_url(location, area, type = "apa")
}
\arguments{
\item{location}{The general location of your search (e.g. "Seattle").  This
is used as the craigslist subdomain, (e.g. "seattle.craigslist.org).}

\item{area}{The specific area of your search (e.g. "tacoma"). This
is used to refine the search.}

\item{type}{What type of thing you want to look up on craiglist.  Currently
only apartment searches are available.  Default is \code{apa} for "apartment".}
}
\value{
base_url A character string containing the base url upon which
queries will be added.
}
\description{
Create the base url based on supplied location data and
desired search type.
}
\keyword{internal}

