% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_num_results-internal.R
\name{get_num_results}
\alias{get_num_results}
\title{get_num_results}
\usage{
get_num_results(query)
}
\arguments{
\item{query}{The URL specifying the query}
}
\value{

}
\description{
Find out how many results are available for the search
}
\keyword{internal}

