% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility-internal.R
\name{na_error}
\alias{na_error}
\title{Convert errors to NA}
\usage{
na_error(expr)
}
\arguments{
\item{expr}{The expression to evaluate}
}
\description{
Given an expression, try to evaluate and return NA if there is
an error.
}
\keyword{internal}

