% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility-internal.R
\name{create_vector}
\alias{create_vector}
\title{Create empty vectors}
\usage{
create_vector(...)
}
\arguments{
\item{...}{Names for any number of vectors to create}
}
\description{
Create any number of empty vectors and assign them to the parent
environment.
}
\keyword{internal}

