% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_housing.R
\name{list_housing}
\alias{list_housing}
\title{List housing}
\usage{
list_housing(location = "Seattle", area = "seattle", base_url = NULL,
  query = NULL, bedrooms = NULL, bathrooms = NULL, min_price = NULL,
  max_price = NULL, min_sqft = NULL, max_sqft = NULL, has_pic = FALSE,
  posted_today = FALSE, pets_cat = FALSE, pets_dog = FALSE)
}
\arguments{
\item{city}{Character string containing the city you want to look in. Default
is "Seattle".}

\item{key}{Optional character string containing a search term to filter
results.}

\item{price_min}{Optional numeric containing minimum price.}

\item{price_max}{Optional numeric containing maximum price.}

\item{sqft_min}{Optional numeric containing minimum square footage.}

\item{sqft_max}{Optional numeric containing maximum square footage.}
}
\description{
Get a list of housing available on craiglist using your own
specifications
}

