% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility-internal.R
\name{na_error}
\alias{na_error}
\title{Errors to NA}
\usage{
na_error(obj)
}
\arguments{
\item{obj}{Object to check for an error}
}
\description{
If an object contains an error, return NA
}
\keyword{internal}

